size_t does not exist for Ardiuno, welcome to hell

+++ bluetooth/nrf52/ChirpBluetoothRxTx.h
@@ -2,6 +2,6 @@
 class ChirpBluetoothRxTx
 {
     public:
-        virtual void send(char[], size_t) = 0;
+        virtual void send(char[], int) = 0;
         virtual char* receive() = 0;
 };
